+++ src/parsing_clickhouse.jl
@@ -113,6 +113,11 @@ function expr_to_sql_clickhouse(expr, sq; from_summarize::Bool)
             return :(replaceRegexpOne($str, $pattern, ""))
         elseif @capture(x, ismissing(a_))
             return  "($(string(a)) IS NULL)"
+        elseif @capture(x, column_[key_])
+            # Convert variables to strings if necessary
+            column_str = string(column)
+            key_str = string(key)
+            return """ CASE WHEN $column_str IS NULL THEN NULL ELSE COALESCE(  LIST_EXTRACT( CASE  WHEN '$key_str' IS NULL THEN NULL ELSE ELEMENT_AT($column_str, '$key_str') END, 1 ), NULL) END ***"""
         # Date extraction functions
         elseif @capture(x, year(a_))
             return "toYear(" * string(a) * ")"
@@ -166,7 +171,12 @@ function expr_to_sql_clickhouse(expr, sq; from_summarize::Bool)
                 return string("match(", column, ", '", pattern_str, "')")
             end
         elseif isa(x, Expr) && x.head == :call && x.args[1] == :n && length(x.args) == 1
-            return "COUNT(*)"
+            if from_summarize
+                return "COUNT(*)"
+            else
+                window_clause = construct_window_clause(sq)
+                return "COUNT(*) $(window_clause)"
+            end
             end
         elseif isa(x, SQLQuery)
             return "(__(" * finalize_query(x) * ")__("