+++ src/parsing_mysql.jl
@@ -113,6 +113,11 @@ function expr_to_sql_mysql(expr, sq; from_summarize::Bool)
             return :(REGEXP_REPLACE($str, $pattern, "", 1, 1))
         elseif @capture(x, ismissing(a_))
             return  "($(string(a)) IS NULL)"
+        elseif @capture(x, column_[key_])
+            # Convert variables to strings if necessary
+            column_str = string(column)
+            key_str = string(key)
+            return """ CASE WHEN $column_str IS NULL THEN NULL ELSE COALESCE(  LIST_EXTRACT( CASE  WHEN '$key_str' IS NULL THEN NULL ELSE ELEMENT_AT($column_str, '$key_str') END, 1 ), NULL) END ***"""
         # Date extraction functions
         elseif @capture(x, year(a_))
             return "EXTRACT(YEAR FROM " * string(a) * ")"
@@ -167,9 +172,14 @@ function expr_to_sql_mysql(expr, sq; from_summarize::Bool)
         elseif isa(x, Expr) && x.head == :call && x.args[1] == :n && length(x.args) == 1
             return "COUNT(*)"
             end
-        elseif isa(x, SQLQuery)
-            return "(__(" * finalize_query(x) * ")__("
-        end
+        elseif isa(x, Expr) && x.head == :call && x.args[1] == :n && length(x.args) == 1
+            if from_summarize
+                return "COUNT(*)"
+            else
+                window_clause = construct_window_clause(sq)
+                return "COUNT(*) $(window_clause)"
+            end
+            end
         return x
     end
 end