+++ src/parsing_oracle.jl
@@ -160,7 +160,12 @@ function expr_to_sql_oracle(expr, sq; from_summarize::Bool)
                 return string("REGEXP_LIKE", column, ", '", pattern_str, "')")
             end
         elseif isa(x, Expr) && x.head == :call && x.args[1] == :n && length(x.args) == 1
-            return "COUNT(*)"
+            if from_summarize
+                return "COUNT(*)"
+            else
+                window_clause = construct_window_clause(sq)
+                return "COUNT(*) $(window_clause)"
+            end
             end
         end
         return x